Update channel sorting to be better. Fix issues with floats and data display in new compendium data source. Add old role listing command. Add category support to create channel commands.

+++ data.js
@@ -186,7 +186,7 @@ function parseProperty(value) {
 	if (typeof value === "object") {
 		return parseObject(value);
 	} else if (!isNaN(value) && value != "") {
-		return parseInt(value, 10);
+		return parseFloat(value, 10);
 	} else {
 		return value;
 	}
